Move some functions to more appropriate places
